+++ src/api/providers/zgsm.ts
@@ -1,3 +1,4 @@
+import * as vscode from "vscode"
 import { Anthropic } from "@anthropic-ai/sdk"
 import OpenAI, { AzureOpenAI } from "openai"
 import axios, { AxiosError } from "axios"
@@ -7,7 +8,6 @@ import {
 	azureOpenAiDefaultApiVersion,
 	ModelInfo,
 	// openAiModelInfoSaneDefaults,
-	zgsmModelInfos,
 } from "../../shared/api"
 import { SingleCompletionHandler } from "../index"
 import { convertToOpenAiMessages } from "../transform/openai-format"
@@ -19,10 +19,12 @@ import { XmlMatcher } from "../../utils/xml-matcher"
 import { DEEP_SEEK_DEFAULT_TEMPERATURE } from "./constants"
 import { createHeaders } from "../../zgsmAuth/zgsmAuthHandler"
 import { defaultZgsmAuthConfig } from "../../zgsmAuth/config"
+import { getWorkspacePath } from "../../utils/path"
+import { getZgsmSelectedModelInfo } from "../../shared/getZgsmSelectedModelInfo"
 
 export const defaultHeaders = {
-	"HTTP-Referer": "https://github.com/RooVetGit/Roo-Cline",
-	"X-Title": "Roo Code",
+	"HTTP-Referer": "https://github.com/zgsm-ai/zgsm",
+	"X-Title": "Shenma",
 }
 
 export interface OpenAiHandlerOptions extends ApiHandlerOptions {}
@@ -33,6 +35,8 @@ const AZURE_AI_INFERENCE_PATH = "/models/chat/completions"
 export class ZgsmHandler extends BaseProvider implements SingleCompletionHandler {
 	protected options: OpenAiHandlerOptions
 	private client: OpenAI
+	private taskId = ""
+	private chatType?: "user" | "system"
 
 	constructor(options: OpenAiHandlerOptions) {
 		super()
@@ -154,7 +158,15 @@ export class ZgsmHandler extends BaseProvider implements SingleCompletionHandler
 
 			const stream = await this.client.chat.completions.create(
 				requestOptions,
-				isAzureAiInference ? { path: AZURE_AI_INFERENCE_PATH } : {},
+				Object.assign(isAzureAiInference ? { path: AZURE_AI_INFERENCE_PATH } : {}, {
+					headers: {
+						...defaultHeaders,
+						"x-quota-identity": this.chatType,
+						"zgsm-task-id": this.taskId,
+						"zgsm-client-id": vscode.env.machineId,
+						"zgsm-project-path": encodeURI(getWorkspacePath()),
+					},
+				}),
 			)
 
 			const matcher = new XmlMatcher(
@@ -234,9 +246,11 @@ export class ZgsmHandler extends BaseProvider implements SingleCompletionHandler
 	}
 
 	override getModel(): { id: string; info: ModelInfo } {
+		const id = `${this.options.zgsmModelId || this.options.zgsmDefaultModelId || defaultModelCache}`
+
 		return {
-			id: `${this.options.zgsmModelId || this.options.zgsmDefaultModelId || defaultModelCache}`,
-			info: this.options.openAiCustomModelInfo || zgsmModelInfos.default,
+			id,
+			info: this.options.openAiCustomModelInfo || getZgsmSelectedModelInfo(id),
 		}
 	}
 
@@ -350,6 +364,18 @@ export class ZgsmHandler extends BaseProvider implements SingleCompletionHandler
 	private _isAzureAiInference(baseUrl?: string): boolean {
 		const urlHost = this._getUrlHost(baseUrl)
 		return urlHost.endsWith(".services.ai.azure.com")
+	}
+
+	setTaskId(taskId: string): void {
+		this.taskId = taskId
+	}
+
+	setChatType(type: "user" | "system"): void {
+		this.chatType = type
+	}
+
+	getChatType() {
+		return this.chatType
 	}
 }